transition with supplementary container view
